backend/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi>=0.115.12",
    "numpy>=2.3.0",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
    "uvicorn>=0.34.3",
    "websocket-client>=1.8.0",
]
